@@ -1,28 +1,98 @@
 <template>
-  <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js App"/>
+  <div class="app">
+    <form @submit.prevent>
+      <h4>Create post</h4>
+      <input
+          v-bind:value="title"
+          @input="title = $event.target.value"
+          class="input"
+          type="text"
+          placeholder="Title"
+      >
+      <input
+          v-bind:value="body"
+          @input="body = $event.target.value"
+          class="input"
+          type="text"
+          placeholder="Description"
+      >
+      <button
+          class="btn"
+          @click="createPost">Create
+      </button>
+    </form>
+    <div class="post" v-for="post in posts" :key="post.id">
+      <div><strong>Title: </strong> {{ post.title }}</div>
+      <div><strong>Description: </strong> {{ post.body }}</div>
+
+    </div>
   </div>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld.vue'
-
 export default {
-  name: 'App',
-  components: {
-    HelloWorld
+  data() {
+    return {
+      posts: [
+        {id: 1, title: 'JavaScript', body: 'Description post'},
+        {id: 2, title: 'TypeScript', body: 'Description post 2'},
+        {id: 3, title: 'React', body: 'Description post 3'}
+      ],
+      title: '',
+      body: ''
+    }
+  },
+  methods: {
+    createPost() {
+      const newPost = {
+        id: Date.now(),
+        title: this.title,
+        body: this.body
+      }
+      this.posts.push(newPost);
+      this.title = '';
+      this.body = ''
+    },
+
   }
 }
 </script>
 
 <style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+.app {
+  padding: 10px;
+}
+
+.post {
+  padding: 15px;
+  border: 2px solid #0636ee;
+  margin-top: 15px;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+}
+
+.input {
+  width: 100%;
+  border: 1px solid #0636ee;
+  padding: 10px 15px;
+  margin-top: 15px;
+}
+
+.btn {
+  align-self: flex-end;
+  margin-top: 15px;
+  padding: 10px 15px;
+  background: none;
+  color: #0636ee;
+  border: #0636ee 1px solid;
 }
 </style>
